refactor dnd logic, to datasets.experimnets reuse existing logic from research metadata
